fix(client): card text is asked for by card, over the English Oracle
The client asked the gateway for card text by printing id, once per game,
and asked Scryfall's /cards/collection about whatever it left out. Text
belongs to the card, not the printing (docs/protocol.md "Card text"), and
the printing key failed a copy outright: a Spark Double's abilities are
printed on the card it copies, whose text the copy's printing does not
carry.

Now the client asks `/catalog/text?lang=..&oracle_ids=..` about the cards
the seat's view names (`PlayerView::cards`, new: the same walk `prints`
takes, plus each object's and stack ability's rules card), each once per
language, one request at a time, 500 at most. Nothing is asked under
English. A gateway that fails, or answers 400 because it predates
`oracle_ids`, is not asked again until the language or the game changes,
and a failed request's cards are asked about again then.

Every face and row draws the compiled English Oracle where the table has
nothing (`CardTexts::face` -> `cardtext::english`; `sentence` pairs the
served `printed` text with the Oracle via `baylee_cardtext::align`, placed
once when filed). An offline duel reads English on every card instead of
blank faces. The Scryfall collection door is deleted: a deck names English
printings, so it could only fetch the English the Oracle already has.

Panels redraw on `CardTexts::generation` rather than a count of entries,
which does not move when a cached entry is replaced by a fresher one.
Old cache entries name no oracle_id and are dropped when read.

Test fixtures file real pool cards with their real German printings (read
from the catalog) instead of invented text under invented indices.

Refs #212

+++ crates/baylee-client/src/hud.rs
@@ -1145,9 +1145,10 @@ pub struct HudRevision {
     /// Whether cards are drawing their constructed face. Held on a key, so
     /// it changes between snapshots and has to be part of the redraw gate.
     faces: bool,
-    /// How many printings have text. Text arrives over the network mid-game,
-    /// and a face built before it lands says a good deal less.
-    texts: usize,
+    /// Which filing of the card-text table this was drawn at. Text arrives
+    /// over the network mid-game, and a face built before it lands says a
+    /// good deal less.
+    texts: u64,
     /// The last refusal, so that one appearing rebuilds the bar.
     ///
     /// Separate from `prompt` because it no longer replaces it: a refusal
